feature: add cct function to trace class
