+++ pipit/readers/hpctoolkit_reader.py
@@ -307,4 +307,4 @@ class HPCToolkitReader:
             by="Time", axis=0, ascending=True, inplace=True, ignore_index=True
         )
         self.trace_df = trace_df
-        return pipit.trace.Trace(None, trace_df)
+        return pipit.trace.Trace(None, trace_df, graph)
